// src/pages/SaleStorage/typings.d.ts
// @ts-ignore
/* eslint-disable */

declare namespace API {
  type Params = {
    beginTime: string,
    customerName: string,
    endTime: string,
    orderNo: string,
    pageNumber: number,
    pageSize: number,
    saleStockOrderType: string
  };

  type TableItem = {
    saleStockOrderId: string,
    customerName: string,
    deliveryAmount: number,
    deliveryTime: string,
    orderNo: string,
    remark: string,
    saleStockOrderNo: string
  }
}